#63 refactor: application 클래스 위치 변경
